requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
frozenlist==1.5.0
futures==3.0.5
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
joblib==1.4.2
multidict==6.1.0
nltk==3.9.1
propcache==0.2.0
python-dotenv==1.0.1
python-telegram-bot==21.6
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.5
urllib3==2.2.3
yarl==1.16.0
